daemon-service/requirements.txt
aiohttp==3.9.1
aioimaplib==1.0.1
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
email-reply-parser==0.5.12
frozenlist==1.4.1
idna==3.6
mail-parser==3.15.0
multidict==6.0.4
simplejson==3.19.2
six==1.16.0
yarl==1.9.4
openai>=1.95.1
mcp==1.11.0
google-api-python-client==2.112.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
flask==2.3.3
flask-cors==4.0.0
python-dateutil==2.8.2
